FIR/UAST: commonize Identifier implementations
