+++ plugins/uast-kotlin-base/src/org/jetbrains/uast/kotlin/declarations/UastLightIdentifier.kt
@@ -1,20 +1,9 @@
 /*
- * Copyright 2010-2017 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.uast.kotlin.declarations
+package org.jetbrains.uast.kotlin
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.psi.PsiElement
@@ -26,12 +15,15 @@ import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.utils.KotlinExceptionWithAttachments
 import org.jetbrains.uast.UElement
 import org.jetbrains.uast.UIdentifier
-import org.jetbrains.uast.kotlin.unwrapFakeFileForLightClass
 import org.jetbrains.uast.toUElement
 
-class UastLightIdentifier(lightOwner: PsiNameIdentifierOwner, ktDeclaration: KtDeclaration?) :
-    KtLightIdentifier(lightOwner, ktDeclaration) {
-    override fun getContainingFile(): PsiFile = unwrapFakeFileForLightClass(super.getContainingFile())
+class UastLightIdentifier(
+    lightOwner: PsiNameIdentifierOwner,
+    ktDeclaration: KtDeclaration?
+) : KtLightIdentifier(lightOwner, ktDeclaration) {
+    override fun getContainingFile(): PsiFile {
+        return unwrapFakeFileForLightClass(super.getContainingFile())
+    }
 }
 
 class KotlinUIdentifier constructor(
